color_server/requirements.txt
Pillow==9.1.0
Flask==2.1.2
Flask_RESTful==0.3.9
marshmallow==3.15.0
flask-cors==3.0.10
pytest==7.1.2